requirements.txt
-e .

Flask==1.1.2
Flask-Cors==3.0.9
numpy==1.19.2
Pillow==7.2.0
scikit-learn==0.23.2
matplotlib==3.3.2
gunicorn==20.0.4